Move HttpResponse to pfm-http crate
