+++ pfm-core/src/forex_manager_impl/mod.rs
@@ -1,2 +1,1 @@
-pub mod forex_manager;
 pub mod forex_manager_storage;
